Optimized portions of peripheral library in analog/serial examples
Analog/Serial examples now only contain the portions of the peripheral
library they require to function (to cut down on maintenance).  These
portions will still be removed once new SDK is published.
