--- Hero Analog Input Example/HERO.IO.cs
@@ -1,105 +1,0 @@
-using Microsoft.SPOT.Hardware;
-
-namespace CTRE
-{
-    namespace HERO
-    {
-        namespace IO
-        {
-            public static class Port1
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x02;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x00;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x01;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x4A;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x4E;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x4D;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x4C;
-
-                public const Cpu.AnalogChannel Analog_Pin3 = Cpu.AnalogChannel.ANALOG_0;
-                public const Cpu.AnalogChannel Analog_Pin4 = Cpu.AnalogChannel.ANALOG_1;
-                public const Cpu.AnalogChannel Analog_Pin5 = Cpu.AnalogChannel.ANALOG_2;
-
-                public const string UART = "COM4";
-            }
-            public static class Port2
-            {
-                //public const Cpu.Pin port2_pin3
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x09;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x0A;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x10;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x07;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x06;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x05;
-            }
-            public static class Port3
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x43;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x45;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x44;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x46;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x14;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x03;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x08;
-            }
-            public static class Port4
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x37;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x35;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x36;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x34;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x33;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x1B;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0xBA;
-
-                public const string UART = "COM2";
-            }
-            public static class Port5
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x30;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x28;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x29;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x32;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x2A;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x2B;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x2C;
-            }
-            public static class Port6
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x41;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x36;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x37;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x40;
-                //public const Cpu.Pin Pin7  Not Connected
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x1B;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x1A;
-
-                public const string UART = "COM6";
-            }
-            public static class Port7
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x22;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x3A;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x3F;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x3B;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x3E;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x3C;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x3D;
-            }
-            public static class Port8
-            {
-                public const Cpu.Pin Pin3 = (Cpu.Pin)0x20;
-                public const Cpu.Pin Pin4 = (Cpu.Pin)0x21;
-                public const Cpu.Pin Pin5 = (Cpu.Pin)0x04;
-                public const Cpu.Pin Pin6 = (Cpu.Pin)0x4B;
-                public const Cpu.Pin Pin7 = (Cpu.Pin)0x4E;
-                public const Cpu.Pin Pin8 = (Cpu.Pin)0x4D;
-                public const Cpu.Pin Pin9 = (Cpu.Pin)0x4C;
-
-                public const Cpu.AnalogChannel Analog_Pin3 = Cpu.AnalogChannel.ANALOG_3;
-                public const Cpu.AnalogChannel Analog_Pin4 = Cpu.AnalogChannel.ANALOG_4;
-                public const Cpu.AnalogChannel Analog_Pin5 = Cpu.AnalogChannel.ANALOG_5;
-            }
-        }
-    }
-}
